requirements.txt
chris_plugin==0.4.0

matplotlib
monai-weekly
gdown
nibabel
tqdm
ignite

torch
pytorch-ignite
itk
scikit-image
scipy
pillow
tensorboard
torchvision
psutil
pandas
lmdb
einops
transformers
mlflow
pynrrd
clearml

